# backend/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Zyvor Fabric API
  description: Zyvor Fabric control plane (zyvor-fabricd daemon) — REST and WebSocket API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:9095
    description: Local development server
paths:
...
  /api/logs:
    get:
      summary: System journal logs
      tags: [Machina v0.1, Observability]
      parameters:
        - $ref: '#/components/parameters/LogLines'
        - $ref: '#/components/parameters/LogPriority'
        - $ref: '#/components/parameters/LogGrep'
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
  /api/vms/{name}/logs:
    get:
      summary: VM journal logs
      tags: [Machina v0.1, Observability]
      parameters:
        - $ref: '#/components/parameters/VmName'
        - $ref: '#/components/parameters/LogLines'
        - $ref: '#/components/parameters/LogPriority'
        - $ref: '#/components/parameters/LogGrep'
      responses:
        '200':
          description: VM log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    VmName:
      name: name
      in: path
      required: true
      schema:
        type: string
    LogLines:
      name: lines
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
    LogPriority:
      name: priority
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 7
    LogGrep:
      name: grep
      in: query
      schema:
        type: string
  schemas:
...
    LogEntry:
      type: object
      properties:
        timestamp:
          type: string
        hostname:
          type: string
        unit:
          type: string
        message:
          type: string
        priority:
          type: string
    LogResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        count:
          type: integer
...
security:
  - bearerAuth: []